update
1) changed long unsigned int to unsigned int for compatibility with jsoncpp
2) createJSON method created
3) changes to opencv manager

+++ src/draw_skeleton_video_3D/Managers/RealSenseManagers/RealSenseManager.cpp
@@ -9,7 +9,7 @@
 
 
 
-void RealSenseManager::setFrameID (long unsigned int frameID) {
+void RealSenseManager::setFrameID (unsigned int frameID) {
     this->frameID = frameID;
 }
 
@@ -55,7 +55,7 @@ RealSenseManager::RealSenseManager (void) : align(RS2_STREAM_COLOR) {
     setFrameID(0);
 }
 
-long unsigned int RealSenseManager::getFrameID (void) {
+unsigned int RealSenseManager::getFrameID (void) {
     return this->frameID;
 }
